Added deleteId and expiry columns. Pastes deleted after expiry.
Added deleteId and expiry columns to the table. Expiry column holds an expiry date after which the paste is deleted. This is done by a recurring task which is executed every X seconds (here 60). Also a delete request should delete the paste with a specific deleteId (provided in the request) which is yet to be implemented.